Add sources to Scaladoc
This uses the `-sourcepath` and `-doc-source-url` options when
generating Scaladoc in add a link to the Chisel3 GitHub source file.
This is setup to link to master if a "-SNAPSHOT" version is used. If
the documentation is built for a non-snapshot version, then this will
use the "v$version" branch on GitHub.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>